requirements.txt
python-dotenv==1.0.0
colorama==0.4.6
qbittorrent-api==2024.12.71